+++ hcloud/resource_hcloud_server_test.go
@@ -17,8 +17,7 @@ import (
 
 var (
 	testAccSSHPublicKey string
-	testHcloudISOID     = "2045"
-	testHcloudISOName   = "coreos_production_iso_image.iso"
+	testHcloudISOID     = "3500"
 )
 
 func init() {
@@ -165,30 +164,6 @@ func TestAccHcloudServer_ISOID(t *testing.T) {
 	})
 }
 
-func TestAccHcloudServer_ISOName(t *testing.T) {
-	var server hcloud.Server
-	rInt := acctest.RandInt()
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccHcloudPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccHcloudCheckServerDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: testAccHcloudCheckServerConfig_ISOName(rInt),
-				Check: resource.ComposeTestCheckFunc(
-					testAccHcloudCheckServerExists("hcloud_server.foobar", &server),
-					testAccHcloudCheckServerAttributes(&server),
-					resource.TestCheckResourceAttr(
-						"hcloud_server.foobar", "name", fmt.Sprintf("foo-%d", rInt)),
-					resource.TestCheckResourceAttr(
-						"hcloud_server.foobar", "iso", testHcloudISOName),
-				),
-			},
-		},
-	})
-}
-
 func testAccHcloudCheckServerDestroy(s *terraform.State) error {
 	client := testAccProvider.Meta().(*hcloud.Client)
 
@@ -322,23 +297,6 @@ resource "hcloud_server" "foobar" {
   iso         = "%s"
   ssh_keys    = ["${hcloud_ssh_key.foobar.id}"]
 }`, rInt, testAccSSHPublicKey, rInt, testHcloudISOID)
-}
-
-func testAccHcloudCheckServerConfig_ISOName(rInt int) string {
-	return fmt.Sprintf(`
-resource "hcloud_ssh_key" "foobar" {
-  name       = "foobar-%d"
-  public_key = "%s"
-}
-resource "hcloud_server" "foobar" {
-  name        = "foo-%d"
-  server_type = "cx11"
-  image       = "debian-9"
-  datacenter  = "fsn1-dc14"
-  backups     = true
-  iso         = "%s"
-  ssh_keys    = ["${hcloud_ssh_key.foobar.id}"]
-}`, rInt, testAccSSHPublicKey, rInt, testHcloudISOName)
 }
 
 func testAccHcloudCheckServerConfig_RenameAndResize(rInt int) string {