+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 BUG FIXES:
 * resource/hcloud_ssh_key: Fix panic when we update labels in SSH keys
+* resource/hcloud_server_network: Fix alias ips ignored on creation of server network
 
 NOTES:
-* This release uses Terraform Plugin SDK v1.6.0.
+* This release uses Terraform Plugin SDK v1.7.0.
 
 ## 1.15.0 (November 11, 2019)